Adapted Context/Fiber 1.72.0

+++ include/memoria/v1/core/linked/common/linked_hash.hpp
@@ -59,7 +59,7 @@ public:
         hash_ = hash_ ^ value;
     }
 
-    uint32_t hash() const {
+    uint64_t hash() const {
         return hash_;
     }
 };
